+++ frontend/cashcompass/lib/widgets/transactions_list_widgets/InterpretedTransaction.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/cupertino.dart';
+
+class InterpretedTransaction {
+  String walletName;
+  IconData categoryIcon;
+  IconData signIcon;
+
+  InterpretedTransaction(
+      {this.walletName = "not set",
+      this.categoryIcon = CupertinoIcons.info,
+      this.signIcon = CupertinoIcons.circle});
+}
