+++ frontend/cashcompass/lib/widgets/transactions_list_widgets/transactions_list.dart
@@ -1,8 +1,7 @@
+import 'package:cashcompass/widgets/transactions_list_widgets/InterpretedTransaction.dart';
 import 'package:cashcompass_hook/src/accounts/active_account/active_account.dart';
 import 'package:cashcompass_hook/src/accounts/passive_account/passive_account.dart';
-import 'package:cashcompass_hook/src/data_storage/accoutmanager.dart';
 import 'package:cashcompass_hook/src/transactions/transactions/transaction.dart';
-import 'package:cashcompass_hook/src/transactions/transactions/transactions_factory.dart';
 import 'package:flutter/cupertino.dart';
 
 class TransactionsList extends StatefulWidget {
@@ -55,7 +54,7 @@ class _TransactionsListState extends State<TransactionsList> {
                 ),
               ),
               ...widget.transactions.map((transaction) {
-                _InterpretedTransaction interpretedTransaction =
+                InterpretedTransaction interpretedTransaction =
                     _interpretTransaction(transaction);
                 return CupertinoListTile(
                   title: Row(
@@ -88,6 +87,7 @@ class _TransactionsListState extends State<TransactionsList> {
                     interpretedTransaction.categoryIcon,
                     color: CupertinoColors.black,
                   ),
+                  onTap: _handleTransactionListTileTapped,
                 );
               }).toList(),
             ],
@@ -97,13 +97,11 @@ class _TransactionsListState extends State<TransactionsList> {
     );
   }
 
-  void handleAddTransaction() {}
-
-  _InterpretedTransaction _interpretTransaction(Transaction transaction) {
+  InterpretedTransaction _interpretTransaction(Transaction transaction) {
     final soll = transaction.soll;
     final haben = transaction.haben;
     var walletName;
-    var categoryIcon;
+    // var categoryIcon;
     var signIcon;
     if (soll.runtimeType == PassiveAccount &&
         haben.runtimeType == PassiveAccount) {
@@ -122,57 +120,16 @@ class _TransactionsListState extends State<TransactionsList> {
     } else {
       throw TypeError();
     }
-    return new _InterpretedTransaction(
-        walletName: walletName, categoryIcon: categoryIcon, signIcon: signIcon);
+    return new InterpretedTransaction(
+      walletName: walletName,
+      // categoryIcon: categoryIcon,
+      signIcon: signIcon,
+    );
   }
-}
 
-class MockTransactions {
-  static Accountmanager manager = Accountmanager();
-  static generateTransactions() {
-    var fac = TransactionsFactory(manager);
-    return [
-      fac
-          .create(
-              amount: 100.00,
-              soll: PassiveAccount("Geschenk", 1),
-              haben: ActiveAcount("ING", 2))
-          .build(),
-      fac
-          .create(
-              amount: 6.50,
-              soll: ActiveAcount("Sprakasse", 3),
-              haben: PassiveAccount("Food", 4))
-          .build(),
-      fac
-          .create(
-              amount: 25.00,
-              soll: ActiveAcount("ING", 5),
-              haben: PassiveAccount("Mobilität", 6))
-          .build(),
-      fac
-          .create(
-              amount: 50.00,
-              soll: PassiveAccount("Geschenk", 7),
-              haben: ActiveAcount("BAR", 8))
-          .build(),
-      fac
-          .create(
-              amount: 450.00,
-              soll: PassiveAccount("Arbeit", 9),
-              haben: ActiveAcount("ING", 10))
-          .build(),
-    ];
+  void handleAddTransaction() {}
+
+  void _handleTransactionListTileTapped() {
+    print('Not yet implemented');
   }
-}
-
-class _InterpretedTransaction {
-  var walletName;
-  var categoryIcon;
-  var signIcon;
-
-  _InterpretedTransaction(
-      {required this.walletName,
-      required this.categoryIcon,
-      required this.signIcon});
 }
